#132 완료 여부에 따른 내 일정 가져오기 메소드 이동 및 PlanService에서 제거
